Remove Footer Component and Refactor Header and Settings Components
- Deleted FooterComponent and its related files
- Updated Header component to remove logout button and company logo
- Refactored Settings component with improved UI and added logout functionality
- Updated color scheme from secondary to slate in various components
- Added company logo color configuration in environment files
