+++ ui/src/app/app.component.html
@@ -10,6 +10,5 @@
       <router-outlet />
     </div>
     <app-toaster />
-    <app-footer />
   </div>
 </div>
