--- ui/src/app/components/layout/footer/footer.component.html
@@ -1,12 +1,0 @@
-<footer *ngIf="authService.isLoggedIn$ | async">
-  <div class="mx-auto px-8 py-3 border-t">
-    <div
-      class="flex items-center justify-center space-x-1 text-center text-xs text-secondary-900 relative">
-      <ng-container *ngIf="themeConfiguration.companyName.length > 0">
-        <span>Accelerate SDLC process with {{ themeConfiguration.companyName }}</span>
-      </ng-container>
-      <span class="px-2 mx-2 font-medium text-xs text-white absolute right-0 bg-primary-400 rounded-full"
-            *ngIf="version.length > 0">{{ version }}</span>
-    </div>
-  </div>
-</footer>
